fix(street_suffix): remove "ga" as a street suffix (#122)
* remove "ga" as a street suffix

this suffix is breaking us queries like "jasper ga"

Wikipedia says "Streets diverging from -gil are named after -gil with the diverging -gil with the additional number in Korean alphabet: ga (가), na (나), da (다), ra (라), ma (마) ... For example, the third diverging street of Daehak-ro 4-gil (대학로4길) would be Daehak-ro 4da-gil (대학로4다길). This secondary diverging numbering is applied to all of the numbering rules. "

But we can't parse the string "Daehak-ro 4da-gil" (or ".. 4ga-gil") so it seems like for now it's better to blacklist this common abbreviation

* add test for jasper ga