fix: Remove unit classifier
This is a temporary fix for issues found parsing some longer inputs.

We don't think the Unit number parsing is especially to blame, it's just
the "last thing" that was added and increases the number of parser
solution combinations.

Since unit number support in Pelias is still in progress, it's a
reasonble thing to remove as a temporary workaround until we have a more
permanent fix.

Connects https://github.com/pelias/api/issues/1530